Migrate to MSBuild project system
Migrate to the MSBuild project system using rc4-004788 of the tooling.
