+++ .editorconfig
@@ -9,5 +9,5 @@ indent_style = space
 insert_final_newline = true
 trim_trailing_whitespace = true
 
-[*.json]
+[*.{csproj,json,props}]
 indent_size = 2
