+++ Build.ps1
@@ -12,8 +12,7 @@ $ErrorActionPreference = "Stop"
 
 $solutionPath  = Split-Path $MyInvocation.MyCommand.Definition
 $framework     = "netcoreapp1.1"
-$getDotNet     = Join-Path $solutionPath "tools\install.ps1"
-$dotnetVersion = "1.0.0-preview2-1-003177"
+$dotnetVersion = "1.0.0-rc4-004788"
 
 if ($OutputPath -eq "") {
     $OutputPath = "$(Convert-Path "$PSScriptRoot")\artifacts"
@@ -28,7 +27,9 @@ if ($env:CI -ne $null -Or $env:TF_BUILD -ne $null) {
 
 if (!(Test-Path $env:DOTNET_INSTALL_DIR)) {
     mkdir $env:DOTNET_INSTALL_DIR | Out-Null
-    & $getDotNet -Version "$dotnetVersion" -InstallDir "$env:DOTNET_INSTALL_DIR" -NoPath
+    $installScript = Join-Path $env:DOTNET_INSTALL_DIR "install.ps1"
+    Invoke-WebRequest "https://raw.githubusercontent.com/dotnet/cli/rel/1.0.0/scripts/obtain/dotnet-install.ps1" -OutFile $installScript
+    & $installScript -Version "$dotnetVersion" -InstallDir "$env:DOTNET_INSTALL_DIR" -NoPath
 }
 
 $env:PATH = "$env:DOTNET_INSTALL_DIR;$env:PATH"
@@ -62,9 +63,9 @@ function DotNetTest { param([string]$Project)
 function DotNetPublish { param([string]$Project)
     $publishPath = (Join-Path $OutputPath "publish")
     if ($VersionSuffix) {
-        & $dotnet publish $Project --output $publishPath --framework $framework --configuration $Configuration --version-suffix "$VersionSuffix" --no-build
+        & $dotnet publish $Project --output $publishPath --framework $framework --configuration $Configuration --version-suffix "$VersionSuffix"
     } else {
-        & $dotnet publish $Project --output $publishPath --framework $framework --configuration $Configuration --no-build
+        & $dotnet publish $Project --output $publishPath --framework $framework --configuration $Configuration
     }
     if ($LASTEXITCODE -ne 0) {
         throw "dotnet publish failed with exit code $LASTEXITCODE"
@@ -89,16 +90,16 @@ if ($PatchVersion -eq $true) {
 }
 
 $projects = @(
-    (Join-Path $solutionPath "src\Website\project.json"),
-    (Join-Path $solutionPath "tests\Website.Tests\project.json")
+    (Join-Path $solutionPath "src\Website\Website.csproj"),
+    (Join-Path $solutionPath "tests\Website.Tests\Website.Tests.csproj")
 )
 
 $testProjects = @(
-    (Join-Path $solutionPath "tests\Website.Tests\project.json")
+    (Join-Path $solutionPath "tests\Website.Tests\Website.Tests.csproj")
 )
 
 $publishProjects = @(
-    (Join-Path $solutionPath "src\Website\project.json")
+    (Join-Path $solutionPath "src\Website\Website.csproj")
 )
 
 if ($RestorePackages -eq $true) {
