+++ Website.Common.props
@@ -1,0 +1,29 @@
+﻿<Project>
+  <ItemGroup>
+    <AdditionalFiles Include="$(MSBuildThisFileDirectory)stylecop.json" Link="stylecop.json" />
+  </ItemGroup>
+  <PropertyGroup>
+    <Authors>Martin Costello</Authors>
+    <CodeAnalysisRuleSet>../../Website.ruleset</CodeAnalysisRuleSet>
+    <Company>https://github.com/martincostello/website</Company>
+    <Copyright>Martin Costello (c) 2016-2017</Copyright>
+    <DefineConstants>$(DefineConstants)</DefineConstants>
+    <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
+    <GenerateDocumentationFile>false</GenerateDocumentationFile>
+    <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
+    <NeutralLanguage>en-US</NeutralLanguage>
+    <PackageIconUrl>https://martincostello.com/favicon.ico</PackageIconUrl>
+    <PackageProjectUrl>https://github.com/martincostello/website</PackageProjectUrl>
+    <PackageLicenseUrl>$(PackageProjectUrl)/blob/master/LICENSE</PackageLicenseUrl>
+    <PackageReleaseNotes>$(PackageProjectUrl)/releases</PackageReleaseNotes>
+    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
+    <PackageTags></PackageTags>
+    <RepositoryType>git</RepositoryType>
+    <RepositoryUrl>$(PackageProjectUrl).git</RepositoryUrl>
+    <RunCodeAnalysis>true</RunCodeAnalysis>
+    <TargetFramework>netcoreapp1.1</TargetFramework>
+    <VersionPrefix>1.0.0</VersionPrefix>
+    <VersionSuffix></VersionSuffix>
+    <WarningsAsErrors>true</WarningsAsErrors>
+  </PropertyGroup>
+</Project>
