+++ Website.sln
@@ -23,11 +23,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{0EED3EF2-37DE-4E6E-B766-3E641D163515}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Website", "src\Website\Website.xproj", "{03AC5596-565F-46B3-A7FD-F70F44C1CBB2}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Website", "src\Website\Website.csproj", "{03AC5596-565F-46B3-A7FD-F70F44C1CBB2}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{668E07A5-C40F-4694-9D3C-EBC1180656CF}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Website.Tests", "tests\Website.Tests\Website.Tests.xproj", "{CB17AD8E-C3F1-45B3-B7BF-E973DAB75840}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Website.Tests", "tests\Website.Tests\Website.Tests.csproj", "{CB17AD8E-C3F1-45B3-B7BF-E973DAB75840}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
