+++ build.sh
@@ -1,5 +1,8 @@
 #!/bin/sh
-dotnet restore --verbosity minimal
-dotnet build src/Website
-dotnet test tests/Website.Tests
-dotnet publish src/Website
+export artifacts=$(dirname "$(readlink -f "$0")")/artifacts
+export configuration=Release
+
+dotnet restore --verbosity minimal || exit 1
+dotnet build --output $artifacts --configuration $configuration || exit 1
+dotnet test tests/Website.Tests/Website.Tests.csproj --output $artifacts --configuration $configuration || exit 1
+dotnet publish src/Website/Website.csproj --output $artifacts --configuration $configuration || exit 1
