+++ global.json
@@ -1,6 +1,5 @@
 ﻿{
-  "projects": [ "src", "tests" ],
   "sdk": {
-    "version": "1.0.0-preview2-003121"
+    "version": "1.0.0-rc4-004788"
   }
 }
