--- src/Website/project.json
@@ -1,131 +1,0 @@
-{
-  "authors": [ "Martin Costello" ],
-  "copyright": "Martin Costello (c) 2016",
-  "description": "https://martincostello.com/",
-  "title": "martincostello.com",
-  "version": "1.0.0-*",
-
-  "buildOptions": {
-    "additionalArguments": [ "/ruleset:../../Website.ruleset", "/additionalfile:../../stylecop.json" ],
-    "emitEntryPoint": true,
-    "preserveCompilationContext": true,
-    "warningsAsErrors": true,
-    "xmlDoc": false,
-    "compile": {
-      "exclude": [ "wwwroot", "node_modules" ],
-      "excludeFiles": [],
-      "includeFiles": [ "../../AssemblyVersion.cs", "../../CommonAssemblyInfo.cs" ]
-    },
-    "copyToOutput": {
-      "includeFiles": [ "appsettings.json", "bower.json", "project.json" ]
-    }
-  },
-
-  "configurations": {
-    "Debug": {
-      "buildOptions": {
-        "define": [ "DEBUG", "TRACE" ],
-        "optimize": false
-      }
-    },
-    "Release": {
-      "buildOptions": {
-        "define": [ "RELEASE", "TRACE" ],
-        "optimize": true
-      }
-    }
-  },
-
-  "dependencies": {
-    "Autofac": "4.3.0",
-    "Autofac.Extensions.DependencyInjection": "4.0.0",
-    "Microsoft.ApplicationInsights.AspNetCore": "2.0.0",
-    "Microsoft.AspNetCore.CookiePolicy": "1.1.0",
-    "Microsoft.AspNetCore.Cors": "1.1.0",
-    "Microsoft.AspNetCore.Diagnostics": "1.1.0",
-    "Microsoft.AspNetCore.HttpOverrides": "1.1.0",
-    "Microsoft.AspNetCore.Mvc": "1.1.1",
-    "Microsoft.AspNetCore.Razor.Tools": {
-      "version": "1.1.0-preview4-final",
-      "type": "build"
-    },
-    "Microsoft.AspNetCore.ResponseCaching": "1.1.0",
-    "Microsoft.AspNetCore.ResponseCompression": "1.0.0",
-    "Microsoft.AspNetCore.Server.IISIntegration": "1.1.0",
-    "Microsoft.AspNetCore.Server.Kestrel": "1.1.0",
-    "Microsoft.AspNetCore.StaticFiles": "1.1.0",
-    "Microsoft.Extensions.Caching.Memory": "1.1.0",
-    "Microsoft.Extensions.Configuration.CommandLine": "1.1.0",
-    "Microsoft.Extensions.Configuration.EnvironmentVariables": "1.1.0",
-    "Microsoft.Extensions.Configuration.Json": "1.1.0",
-    "Microsoft.Extensions.Configuration.UserSecrets": "1.1.0",
-    "Microsoft.Extensions.Logging": "1.1.0",
-    "Microsoft.Extensions.Logging.Console": "1.1.0",
-    "Microsoft.Extensions.Logging.Debug": "1.1.0",
-    "Microsoft.Extensions.Logging.TraceSource": "1.1.0",
-    "Microsoft.Extensions.Options.ConfigurationExtensions": "1.1.0",
-    "Microsoft.VisualStudio.Web.BrowserLink.Loader": "14.1.0",
-    "Newtonsoft.Json": "9.0.1",
-    "NodaTime": "2.0.0-beta20170123",
-    "StyleCop.Analyzers": {
-      "version": "1.0.0",
-      "type": "build"
-    }
-  },
-
-  "packOptions": {
-    "iconUrl": "https://martincostello.com/favicon.ico",
-    "licenseUrl": "https://github.com/martincostello/website/blob/master/LICENSE",
-    "owners": [ "Martin Costello" ],
-    "projectUrl": "https://github.com/martincostello/website",
-    "releaseNotes": "https://github.com/martincostello/website/releases",
-    "repository": {
-      "type": "git",
-      "url": "https://github.com/martincostello/website.git"
-    },
-    "requireLicenseAcceptance": false,
-    "summary": "Martin Costello's website",
-    "tags": []
-  },
-
-  "runtimeOptions": {
-    "gcConcurrent": false,
-    "gcServer": true
-  },
-
-  "tooling": {
-    "defaultNamespace": "MartinCostello.Website"
-  },
-
-  "tools": {
-    "Microsoft.AspNetCore.Razor.Tools": "1.1.0-preview4-final",
-    "Microsoft.Extensions.SecretManager.Tools": "1.1.0-preview4-final",
-    "Microsoft.AspNetCore.Server.IISIntegration.Tools": "1.1.0-preview4-final"
-  },
-
-  "frameworks": {
-    "netcoreapp1.1": {
-      "dependencies": {
-        "Microsoft.NETCore.App": {
-          "version": "1.1.0",
-          "type": "platform"
-        }
-      },
-      "imports": [
-        "dotnet5.6",
-        "portable-net45+win8"
-      ]
-    }
-  },
-
-  "publishOptions": {
-    "include": [ "wwwroot", "Views" ],
-    "includeFiles": [ "appsettings.json", "bower.json", "parameters.xml", "project.json", "web.config" ]
-  },
-
-  "scripts": {
-    "prebuild": [ "gulp build" ],
-    "prepublish": [ "npm install --loglevel=error", "bower install --loglevel=error", "gulp publish" ],
-    "postpublish": [ "dotnet publish-iis --publish-folder %publish:OutputPath% --framework %publish:FullTargetFramework%" ]
-  }
-}
