+++ tests/Website.Tests/Website.Tests.csproj
@@ -1,0 +1,27 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <Import Project="..\..\Website.Common.props" />
+  <PropertyGroup>
+    <AssemblyName>Website.Tests</AssemblyName>
+    <AssemblyTitle>martincostello.com.Tests</AssemblyTitle>
+    <Description>Tests for https://martincostello.com/</Description>
+    <PackageId>Website.Tests</PackageId>
+    <Summary>Tests for Martin Costello's website</Summary>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="..\..\AssemblyVersion.cs;..\..\CommonAssemblyInfo.cs" />
+    <Content Include="testsettings.json;xunit.runner.json" CopyToOutputDirectory="PreserveNewest" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Website\Website.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="1.1.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.0.0-preview-20170106-08" />
+    <PackageReference Include="NodaTime" Version="2.0.0-beta20170123" />
+    <PackageReference Include="NodaTime.Testing" Version="2.0.0-beta20170123" />
+    <PackageReference Include="Shouldly" Version="2.8.2" />
+    <PackageReference Include="StyleCop.Analyzers" Version="1.0.0" PrivateAssets="All" />
+    <PackageReference Include="xunit" Version="2.2.0-beta6-build3495" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.2.0-beta6-build1242" />
+  </ItemGroup>
+</Project>
