--- tests/Website.Tests/project.json
@@ -1,63 +1,0 @@
-﻿{
-  "authors": [ "Martin Costello" ],
-  "copyright": "Martin Costello (c) 2016",
-  "description": "Tests for https://martincostello.com/",
-  "title": "martincostello.com.Tests",
-  "version": "1.0.0-*",
-
-  "buildOptions": {
-    "additionalArguments": [ "/ruleset:../../Website.ruleset", "/additionalfile:../../stylecop.json" ],
-    "compile": {
-      "includeFiles": [ "../../AssemblyVersion.cs", "../../CommonAssemblyInfo.cs" ]
-    },
-    "copyToOutput": {
-      "include": [ "appsettings.json", "xunit.runner.json" ]
-    },
-    "warningsAsErrors": true,
-    "xmlDoc": false
-  },
-
-  "dependencies": {
-    "dotnet-test-xunit": "2.2.0-preview2-build1029",
-    "Microsoft.AspNetCore.TestHost": "1.1.0",
-    "Microsoft.DotNet.InternalAbstractions": "1.0.0",
-    "NodaTime": "2.0.0-beta20170123",
-    "NodaTime.Testing": "2.0.0-beta20170123",
-    "Shouldly": "2.8.2",
-    "StyleCop.Analyzers": {
-      "version": "1.0.0",
-      "type": "build"
-    },
-    "Website": "1.0.0-*",
-    "xunit": "2.2.0-beta4-build3444"
-  },
-
-  "frameworks": {
-    "netcoreapp1.1": {
-      "dependencies": {
-        "Microsoft.NETCore.App": {
-          "type": "platform",
-          "version": "1.1.0"
-        }
-      },
-      "imports": "dotnet5.6"
-    }
-  },
-
-  "packOptions": {
-    "iconUrl": "https://martincostello.com/favicon.ico",
-    "licenseUrl": "https://github.com/martincostello/website/blob/master/LICENSE",
-    "owners": [ "Martin Costello" ],
-    "projectUrl": "https://github.com/martincostello/website",
-    "releaseNotes": "https://github.com/martincostello/website/releases",
-    "repository": {
-      "type": "git",
-      "url": "https://github.com/martincostello/website.git"
-    },
-    "requireLicenseAcceptance": false,
-    "summary": "Tests for Martin Costello's website",
-    "tags": []
-  },
-
-  "testRunner": "xunit"
-}
